refactor(position-keeping): extract log endpoints from server.go
Move RetrieveFinancialPositionLog and ListFinancialPositionLogs RPC
handlers to grpc_log_endpoints.go. Move fromProtoTransactionStatus
mapper to adapters.go alongside existing proto conversion functions.

server.go now contains only the service struct, constructor, options,
constants, and error definitions.
